refactored GetFiles to use a DAO
