+++ src/main/java/com/photosharesite/backend/PhotoShareSiteApplication.java
@@ -1,6 +1,7 @@
 package com.photosharesite.backend;
 
 import com.photosharesite.backend.db.insertorselectuser.InsertOrSelectUserAccess;
+import com.photosharesite.backend.db.selectfiles.SelectFilesAccess;
 import com.photosharesite.backend.endpoints.getfiles.GetFilesResource;
 import com.photosharesite.backend.resources.HelloWorldResource;
 import com.photosharesite.backend.endpoints.lookupuser.LookupUserResource;
@@ -57,7 +58,9 @@ public class PhotoShareSiteApplication extends Application<PhotoShareSiteConfigu
         environment.jersey().register(lookupUserResource);
 
         // create and register GetFilesResource
-        final GetFilesResource getFilesResource = new GetFilesResource(jdbi);
+        final GetFilesResource getFilesResource = new GetFilesResource(
+                new SelectFilesAccess(jdbi)
+        );
         environment.jersey().register(getFilesResource);
     }
     private void initSwagger(PhotoShareSiteConfiguration configuration, Environment environment) {
